src/trade_execution.py
from schwabdev.api import TradingAPI
import joblib
import pandas as pd

def execute_trades(api, model, stock_data):
    signals = generate_signals(model, stock_data)
    for symbol, signal in signals.items():
        if signal == "buy":
            api.place_order("YOUR_ACCOUNT_ID", {"symbol": symbol, "action": "BUY", "quantity": 10})
        elif signal == "sell":
            api.place_order("YOUR_ACCOUNT_ID", {"symbol": symbol, "action": "SELL", "quantity": 10})

def generate_signals(model, stock_data):
    signals = {}
    for symbol, df in stock_data.items():
        X = df[['MA10', 'MA50', 'RSI', 'MACD', 'BB_high', 'BB_low']]
        y_pred = model.predict(X)
        signals[symbol] = "buy" if y_pred[-1] == 1 else "sell"
    return signals
