requirements.txt
asyncio==3.4.3
matplotlib==3.6.2
numpy
pandas==1.5.3
python-dotenv
requests==2.28.1
schedule==1.1.0
selenium
webdriver_manager
websockets==10.3
yfinance
ta
xgboost
scikit-learn
joblib
sys
json